refactor: List<section>을 일급컬렉션 Sections으로 변경
